+++ src/Position.cpp
@@ -29,7 +29,7 @@ void Position::to_struct(PositionStruct &position_struct){
 	position_struct.unrealized_pl = this->unrealized_pl;
 }
 
-Position::Position(unsigned int position_id, unsigned int asset_id, float units, float average_price, timeval position_create_time,
+Position::Position(unsigned int position_id, unsigned int asset_id, double units, double average_price, timeval position_create_time,
 			unsigned int exchange_id,
 			unsigned int account_id,
 			unsigned int strategy_id) {
@@ -43,24 +43,24 @@ Position::Position(unsigned int position_id, unsigned int asset_id, float units,
 	this->account_id = account_id;
 	this->strategy_id = strategy_id;
 }
-void Position::increase(float market_price, float _units) {
-	float new_units = abs(this->units) + abs(_units);
+void Position::increase(double market_price, double _units) {
+	double new_units = abs(this->units) + abs(_units);
 	this->average_price = ((abs(this->units)*this->average_price) + (abs(_units)*market_price)) / new_units;
 	this->units += _units;
 	this->bars_since_change = 0;
 }
-void Position::reduce(float market_price, float _units) {
+void Position::reduce(double market_price, double _units) {
 	this->realized_pl += abs(_units) * (market_price - this->average_price);
 	this->units -= abs(_units);
 	this->bars_since_change = 0;
 }
-void Position::close(float close_price, timeval position_close_time) {
+void Position::close(double close_price, timeval position_close_time) {
 	this->is_open = false;
 	this->close_price = close_price;
 	this->position_close_time = position_close_time;
 	this->realized_pl += this->units * (close_price - this->average_price);
 	this->unrealized_pl = 0;
 }
-float Position::liquidation_value() {
+double Position::liquidation_value() {
 	return this->units * this->last_price;
 }
